[FEAT] Thiết kế lại hệ thống model cho product

+++ SaleCar/src/main/java/code/salecar/dao/ImageDAO.java
@@ -12,12 +12,12 @@ import java.util.List;
 
 public class ImageDAO {
 
-    public String getImage(Image.entityType type, int id) {
+    public String getImage(Image.entityType type, long id) {
         String query = "select image_url1 from image where entity_type = ? and entity_id=?";
         try (Connection con = DBConnection.getConnection();
              PreparedStatement ps = con.prepareStatement(query);) {
             ps.setString(1, type.toString());
-            ps.setInt(2, id);
+            ps.setLong(2, id);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
                 return rs.getString("image_url1");
@@ -32,13 +32,13 @@ public class ImageDAO {
 
     }
 
-    public List<String> getImageProduct(int id) {
+    public List<String> getImageProduct(long id) {
         List<String> list = new ArrayList<>();
         String query = "select image_url1 from image where entity_type = ? and entity_id=?";
         try (Connection con = DBConnection.getConnection();
              PreparedStatement ps = con.prepareStatement(query);) {
             ps.setString(1, Image.entityType.product.toString());
-            ps.setInt(2, id);
+            ps.setLong(2, id);
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
                 list.add(rs.getString("image_url1"));
